+++ pdata/pmetric/generated_metrics.go
@@ -21,6 +21,7 @@ import (
 	"sort"
 
 	"go.opentelemetry.io/collector/pdata/internal"
+	"go.opentelemetry.io/collector/pdata/internal/data"
 	otlpmetrics "go.opentelemetry.io/collector/pdata/internal/data/protogen/metrics/v1"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 )
@@ -1241,13 +1242,13 @@ func (ms NumberDataPoint) Exemplars() ExemplarSlice {
 	return ExemplarSlice(internal.NewExemplarSlice(&ms.getOrig().Exemplars))
 }
 
-// FlagsImmutable returns the flagsimmutable associated with this NumberDataPoint.
-func (ms NumberDataPoint) FlagsImmutable() MetricDataPointFlagsImmutable {
-	return MetricDataPointFlagsImmutable(ms.getOrig().Flags)
+// Flags returns the flags associated with this NumberDataPoint.
+func (ms NumberDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.getOrig().Flags)
 }
 
-// SetFlagsImmutable replaces the flagsimmutable associated with this NumberDataPoint.
-func (ms NumberDataPoint) SetFlagsImmutable(v MetricDataPointFlagsImmutable) {
+// SetFlags replaces the flags associated with this NumberDataPoint.
+func (ms NumberDataPoint) SetFlags(v MetricDataPointFlags) {
 	ms.getOrig().Flags = uint32(v)
 }
 
@@ -1264,7 +1265,7 @@ func (ms NumberDataPoint) CopyTo(dest NumberDataPoint) {
 	}
 
 	ms.Exemplars().CopyTo(dest.Exemplars())
-	dest.SetFlagsImmutable(ms.FlagsImmutable())
+	dest.SetFlags(ms.Flags())
 }
 
 // HistogramDataPointSlice logically represents a slice of HistogramDataPoint.
@@ -1516,13 +1517,13 @@ func (ms HistogramDataPoint) Exemplars() ExemplarSlice {
 	return ExemplarSlice(internal.NewExemplarSlice(&ms.getOrig().Exemplars))
 }
 
-// FlagsImmutable returns the flagsimmutable associated with this HistogramDataPoint.
-func (ms HistogramDataPoint) FlagsImmutable() MetricDataPointFlagsImmutable {
-	return MetricDataPointFlagsImmutable(ms.getOrig().Flags)
+// Flags returns the flags associated with this HistogramDataPoint.
+func (ms HistogramDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.getOrig().Flags)
 }
 
-// SetFlagsImmutable replaces the flagsimmutable associated with this HistogramDataPoint.
-func (ms HistogramDataPoint) SetFlagsImmutable(v MetricDataPointFlagsImmutable) {
+// SetFlags replaces the flags associated with this HistogramDataPoint.
+func (ms HistogramDataPoint) SetFlags(v MetricDataPointFlags) {
 	ms.getOrig().Flags = uint32(v)
 }
 
@@ -1583,7 +1584,7 @@ func (ms HistogramDataPoint) CopyTo(dest HistogramDataPoint) {
 	}
 
 	ms.Exemplars().CopyTo(dest.Exemplars())
-	dest.SetFlagsImmutable(ms.FlagsImmutable())
+	dest.SetFlags(ms.Flags())
 	if ms.HasMin() {
 		dest.SetMin(ms.Min())
 	}
@@ -1856,13 +1857,13 @@ func (ms ExponentialHistogramDataPoint) Exemplars() ExemplarSlice {
 	return ExemplarSlice(internal.NewExemplarSlice(&ms.getOrig().Exemplars))
 }
 
-// FlagsImmutable returns the flagsimmutable associated with this ExponentialHistogramDataPoint.
-func (ms ExponentialHistogramDataPoint) FlagsImmutable() MetricDataPointFlagsImmutable {
-	return MetricDataPointFlagsImmutable(ms.getOrig().Flags)
+// Flags returns the flags associated with this ExponentialHistogramDataPoint.
+func (ms ExponentialHistogramDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.getOrig().Flags)
 }
 
-// SetFlagsImmutable replaces the flagsimmutable associated with this ExponentialHistogramDataPoint.
-func (ms ExponentialHistogramDataPoint) SetFlagsImmutable(v MetricDataPointFlagsImmutable) {
+// SetFlags replaces the flags associated with this ExponentialHistogramDataPoint.
+func (ms ExponentialHistogramDataPoint) SetFlags(v MetricDataPointFlags) {
 	ms.getOrig().Flags = uint32(v)
 }
 
@@ -1913,7 +1914,7 @@ func (ms ExponentialHistogramDataPoint) CopyTo(dest ExponentialHistogramDataPoin
 	ms.Positive().CopyTo(dest.Positive())
 	ms.Negative().CopyTo(dest.Negative())
 	ms.Exemplars().CopyTo(dest.Exemplars())
-	dest.SetFlagsImmutable(ms.FlagsImmutable())
+	dest.SetFlags(ms.Flags())
 	if ms.HasMin() {
 		dest.SetMin(ms.Min())
 	}
@@ -2212,13 +2213,13 @@ func (ms SummaryDataPoint) QuantileValues() ValueAtQuantileSlice {
 	return ValueAtQuantileSlice(internal.NewValueAtQuantileSlice(&ms.getOrig().QuantileValues))
 }
 
-// FlagsImmutable returns the flagsimmutable associated with this SummaryDataPoint.
-func (ms SummaryDataPoint) FlagsImmutable() MetricDataPointFlagsImmutable {
-	return MetricDataPointFlagsImmutable(ms.getOrig().Flags)
+// Flags returns the flags associated with this SummaryDataPoint.
+func (ms SummaryDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.getOrig().Flags)
 }
 
-// SetFlagsImmutable replaces the flagsimmutable associated with this SummaryDataPoint.
-func (ms SummaryDataPoint) SetFlagsImmutable(v MetricDataPointFlagsImmutable) {
+// SetFlags replaces the flags associated with this SummaryDataPoint.
+func (ms SummaryDataPoint) SetFlags(v MetricDataPointFlags) {
 	ms.getOrig().Flags = uint32(v)
 }
 
@@ -2230,7 +2231,7 @@ func (ms SummaryDataPoint) CopyTo(dest SummaryDataPoint) {
 	dest.SetCount(ms.Count())
 	dest.SetSum(ms.Sum())
 	ms.QuantileValues().CopyTo(dest.QuantileValues())
-	dest.SetFlagsImmutable(ms.FlagsImmutable())
+	dest.SetFlags(ms.Flags())
 }
 
 // ValueAtQuantileSlice logically represents a slice of ValueAtQuantile.
@@ -2641,22 +2642,22 @@ func (ms Exemplar) FilteredAttributes() pcommon.Map {
 
 // TraceID returns the traceid associated with this Exemplar.
 func (ms Exemplar) TraceID() pcommon.TraceID {
-	return pcommon.NewTraceID(ms.getOrig().TraceId)
+	return pcommon.TraceID(ms.getOrig().TraceId)
 }
 
 // SetTraceID replaces the traceid associated with this Exemplar.
 func (ms Exemplar) SetTraceID(v pcommon.TraceID) {
-	ms.getOrig().TraceId = v.Bytes()
+	ms.getOrig().TraceId = data.TraceID(v)
 }
 
 // SpanID returns the spanid associated with this Exemplar.
 func (ms Exemplar) SpanID() pcommon.SpanID {
-	return pcommon.NewSpanID(ms.getOrig().SpanId)
+	return pcommon.SpanID(ms.getOrig().SpanId)
 }
 
 // SetSpanID replaces the spanid associated with this Exemplar.
 func (ms Exemplar) SetSpanID(v pcommon.SpanID) {
-	ms.getOrig().SpanId = v.Bytes()
+	ms.getOrig().SpanId = data.SpanID(v)
 }
 
 // CopyTo copies all properties from the current struct to the dest.