+++ pdata/pmetric/generated_metrics_test.go
@@ -898,12 +898,12 @@ func TestNumberDataPoint_Exemplars(t *testing.T) {
 	assert.Equal(t, ExemplarSlice(internal.GenerateTestExemplarSlice()), ms.Exemplars())
 }
 
-func TestNumberDataPoint_FlagsImmutable(t *testing.T) {
+func TestNumberDataPoint_Flags(t *testing.T) {
 	ms := NewNumberDataPoint()
-	assert.Equal(t, MetricDataPointFlagsImmutable(0), ms.FlagsImmutable())
-	testValFlagsImmutable := MetricDataPointFlagsImmutable(1)
-	ms.SetFlagsImmutable(testValFlagsImmutable)
-	assert.Equal(t, testValFlagsImmutable, ms.FlagsImmutable())
+	assert.Equal(t, MetricDataPointFlags(0), ms.Flags())
+	testValFlags := MetricDataPointFlags(1)
+	ms.SetFlags(testValFlags)
+	assert.Equal(t, testValFlags, ms.Flags())
 }
 
 func TestHistogramDataPointSlice(t *testing.T) {
@@ -1094,12 +1094,12 @@ func TestHistogramDataPoint_Exemplars(t *testing.T) {
 	assert.Equal(t, ExemplarSlice(internal.GenerateTestExemplarSlice()), ms.Exemplars())
 }
 
-func TestHistogramDataPoint_FlagsImmutable(t *testing.T) {
+func TestHistogramDataPoint_Flags(t *testing.T) {
 	ms := NewHistogramDataPoint()
-	assert.Equal(t, MetricDataPointFlagsImmutable(0), ms.FlagsImmutable())
-	testValFlagsImmutable := MetricDataPointFlagsImmutable(1)
-	ms.SetFlagsImmutable(testValFlagsImmutable)
-	assert.Equal(t, testValFlagsImmutable, ms.FlagsImmutable())
+	assert.Equal(t, MetricDataPointFlags(0), ms.Flags())
+	testValFlags := MetricDataPointFlags(1)
+	ms.SetFlags(testValFlags)
+	assert.Equal(t, testValFlags, ms.Flags())
 }
 
 func TestHistogramDataPoint_Min(t *testing.T) {
@@ -1314,12 +1314,12 @@ func TestExponentialHistogramDataPoint_Exemplars(t *testing.T) {
 	assert.Equal(t, ExemplarSlice(internal.GenerateTestExemplarSlice()), ms.Exemplars())
 }
 
-func TestExponentialHistogramDataPoint_FlagsImmutable(t *testing.T) {
+func TestExponentialHistogramDataPoint_Flags(t *testing.T) {
 	ms := NewExponentialHistogramDataPoint()
-	assert.Equal(t, MetricDataPointFlagsImmutable(0), ms.FlagsImmutable())
-	testValFlagsImmutable := MetricDataPointFlagsImmutable(1)
-	ms.SetFlagsImmutable(testValFlagsImmutable)
-	assert.Equal(t, testValFlagsImmutable, ms.FlagsImmutable())
+	assert.Equal(t, MetricDataPointFlags(0), ms.Flags())
+	testValFlags := MetricDataPointFlags(1)
+	ms.SetFlags(testValFlags)
+	assert.Equal(t, testValFlags, ms.Flags())
 }
 
 func TestExponentialHistogramDataPoint_Min(t *testing.T) {
@@ -1541,12 +1541,12 @@ func TestSummaryDataPoint_QuantileValues(t *testing.T) {
 	assert.Equal(t, ValueAtQuantileSlice(internal.GenerateTestValueAtQuantileSlice()), ms.QuantileValues())
 }
 
-func TestSummaryDataPoint_FlagsImmutable(t *testing.T) {
+func TestSummaryDataPoint_Flags(t *testing.T) {
 	ms := NewSummaryDataPoint()
-	assert.Equal(t, MetricDataPointFlagsImmutable(0), ms.FlagsImmutable())
-	testValFlagsImmutable := MetricDataPointFlagsImmutable(1)
-	ms.SetFlagsImmutable(testValFlagsImmutable)
-	assert.Equal(t, testValFlagsImmutable, ms.FlagsImmutable())
+	assert.Equal(t, MetricDataPointFlags(0), ms.Flags())
+	testValFlags := MetricDataPointFlags(1)
+	ms.SetFlags(testValFlags)
+	assert.Equal(t, testValFlags, ms.Flags())
 }
 
 func TestValueAtQuantileSlice(t *testing.T) {
@@ -1848,16 +1848,16 @@ func TestExemplar_FilteredAttributes(t *testing.T) {
 
 func TestExemplar_TraceID(t *testing.T) {
 	ms := NewExemplar()
-	assert.Equal(t, pcommon.NewTraceID(data.NewTraceID([16]byte{})), ms.TraceID())
-	testValTraceID := pcommon.NewTraceID(data.NewTraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1}))
+	assert.Equal(t, pcommon.TraceID(data.NewTraceID([16]byte{})), ms.TraceID())
+	testValTraceID := pcommon.TraceID(data.NewTraceID([16]byte{1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 6, 5, 4, 3, 2, 1}))
 	ms.SetTraceID(testValTraceID)
 	assert.Equal(t, testValTraceID, ms.TraceID())
 }
 
 func TestExemplar_SpanID(t *testing.T) {
 	ms := NewExemplar()
-	assert.Equal(t, pcommon.NewSpanID(data.NewSpanID([8]byte{})), ms.SpanID())
-	testValSpanID := pcommon.NewSpanID(data.NewSpanID([8]byte{1, 2, 3, 4, 5, 6, 7, 8}))
+	assert.Equal(t, pcommon.SpanID(data.NewSpanID([8]byte{})), ms.SpanID())
+	testValSpanID := pcommon.SpanID(data.NewSpanID([8]byte{8, 7, 6, 5, 4, 3, 2, 1}))
 	ms.SetSpanID(testValSpanID)
 	assert.Equal(t, testValSpanID, ms.SpanID())
 }